Return 32-bit array from assembleBlock

+++ src/assembler.ts
@@ -759,5 +759,5 @@ export const assembleBlock = (asmLines: string[]) => {
         machineCode.push(0);
     }
 
-    return new Uint16Array(machineCode);
+    return new Uint32Array(machineCode);
 };
